Добавил ДТО Сотрудникам и Начал ОРДЕР

+++ Anti-Cafe/src/main/java/com/naumen/anticafe/service/Role/RoleService.java
@@ -7,5 +7,5 @@ import java.util.List;
 
 public interface RoleService {
     List<Role> getAllRole();
-    Role getRole(Long roleId) throws NotFoundException;
+    Role getRole(Integer roleId) throws NotFoundException;
 }
